requirements.txt
babel==2.17.0
flake8==7.1.2
flake8-helper==0.2.2
flake8-strftime==0.3.4
mccabe==0.7.0
mysql-connector==2.2.9
numpy==2.2.3
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
pillow==11.1.0
pycodestyle==2.12.1
pyflakes==3.2.0
PyMySQL==1.1.1
tkcalendar==1.6.1
